# backend/content-service/api/openapi.yaml
openapi: 3.0.3
info:
  title: Verita Content Service API
  description: |
    Comprehensive service for managing the lifecycle of posts, hierarchical comments,
    adding topic, user interactions (likes/bookmarks), and content moderation.
  version: 1.1.0
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: 'http://api.verita.local/content'
    description: Production-style internal URL

paths:
  # --- POSTS MANAGEMENT ---
  /api/v1/posts:
    get:
      operationId: getAllPosts
      summary: Retrieve a paginated list of posts
      description: |
        General-purpose endpoint for fetching a batch of posts with optional topic filtering.
        Intended for use cases such as topic-based browsing, admin tooling, or any client
        that needs raw paginated post data.

        **Not** the homepage feed — the personalised and trending feeds are served by the
        Recommendation Service (`GET /api/v1/feed/personal` and `GET /api/v1/feed/trending`).
        Those endpoints apply algorithmic ranking and return enriched post payloads.

        The `topic` parameter performs an exact-match filter and returns results in
        reverse-chronological order. For popularity-ranked topic browsing, use
        `GET /api/v1/feed/trending?topic=` in the Recommendation Service instead.
      tags: [Posts]
      security: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 0 }
        - name: size
          in: query
          schema: { type: integer, default: 10 }
        - name: topic
          in: query
          schema: { type: string }
      responses:
        '200':
          description: A paginated list of posts retrieved successfully
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PostPage' }
        '400': { $ref: '#/components/responses/BadRequest' }

    post:
      operationId: createPost
      summary: Create a new post or draft
      tags: [Posts]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PostRequest' }
      responses:
        '201':
          description: Post created successfully
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PostResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  /api/v1/posts/cards:
    get:
      operationId: getPostCards
      summary: Fetch lightweight PostCard objects for multiple posts by ID
      description: |
        Returns `PostCard` objects (no article body) for the given post IDs, preserving
        the order of the `ids` parameter. Called by the frontend after receiving a `postIds`
        list from the Recommendation Service to populate feed cards.
      tags: [Posts]
      security: []
      parameters:
        - name: ids
          in: query
          required: true
          description: Comma-separated list of post UUIDs, e.g. `ids=id1,id2,id3`.
          schema:
            type: array
            items: { type: string, format: uuid }
            maxItems: 50
          style: form
          explode: false
      responses:
        '200':
          description: Posts retrieved successfully. Missing IDs are silently omitted.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PostCard' }
        '400': { $ref: '#/components/responses/BadRequest' }

  /api/v1/posts/search:
    get:
      operationId: searchPosts
      summary: Search posts by keyword
      description: |
        Full-text keyword search across post titles and content.
        Returns results ranked by relevance to the query string, not by time or popularity.

        This endpoint is driven by user intent (the user typed something to find) and is
        distinct from feed discovery. Future iterations may back this with a dedicated
        search engine (e.g. Elasticsearch) without changing the contract.

        Authentication is optional. When a valid Bearer token is present, `isLikedByMe`
        and `isBookmarkedByMe` are populated per the authenticated user; otherwise they
        default to `false`.
      tags: [Posts]
      security: []
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string, minLength: 1 }
        - name: page
          in: query
          schema: { type: integer, default: 0 }
        - name: size
          in: query
          schema: { type: integer, default: 10 }
      responses:
        '200':
          description: Search results retrieved successfully
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PostPage' }
        '400': { $ref: '#/components/responses/BadRequest' }

  /api/v1/posts/{id}:
    get:
      operationId: getPostById
      summary: Get detailed post by ID
      tags: [Posts]
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Post details retrieved successfully
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PostResponse' }
        '404': { $ref: '#/components/responses/NotFound' }

    put:
      operationId: updatePost
      summary: Update an existing post
      tags: [Posts]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PostRequest' }
      responses:
        '200':
          description: Post updated successfully
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PostResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

    patch:
      operationId: patchPost
      summary: Partially update an existing post
      description: |
        Updates only the fields provided in the request body; omitted fields are left
        unchanged. Use this for single-field updates such as unpublishing a post
        (`{"status": "DRAFT"}`) without re-sending the full post payload.
        For full replacement (e.g. the post editor), use `PUT`.

        Only the post author may patch a post.
      tags: [Posts]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PostPatchRequest' }
      responses:
        '200':
          description: Post updated successfully
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PostResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

    delete:
      operationId: deletePost
      summary: Soft-delete a post
      tags: [Posts]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: Post deleted successfully
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # --- LIKES ---
  /api/v1/posts/{id}/like:
    post:
      operationId: likePost
      summary: Like or dislike a post
      description: Send LIKE or DISLIKE to cast/change a reaction. Send NONE to remove any existing reaction.
      tags: [Interactions]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LikeRequest' }
      responses:
        '200':
          description: Like processed successfully
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PostLikeResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }

  # --- BOOKMARKS ---
  /api/v1/posts/{id}/bookmark:
    post:
      operationId: bookmarkPost
      summary: Save a post to the current user's bookmarks
      tags: [Interactions]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: Post bookmarked successfully
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }

    delete:
      operationId: unbookmarkPost
      summary: Remove a post from the current user's bookmarks
      tags: [Interactions]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: Post removed from bookmarks successfully
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }

  # --- COMMENTS ---
  /api/v1/posts/{id}/comments:
    get:
      operationId: getCommentsByPost
      summary: Get all comments for a post
      tags: [Comments]
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Hierarchical list of comments retrieved
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CommentResponse' }
        '404': { $ref: '#/components/responses/NotFound' }

    post:
      operationId: createComment
      summary: Post a comment or reply
      tags: [Comments]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CommentRequest' }
      responses:
        '201':
          description: Comment created successfully
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CommentResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  /api/v1/comments/{id}:
    delete:
      operationId: deleteComment
      summary: Delete a comment owned by the current user
      tags: [Comments]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: Comment deleted successfully
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  /api/v1/comments/{id}/like:
    post:
      operationId: likeComment
      summary: Like a comment
      tags: [Interactions]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Comment like processed successfully
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CommentLikeResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }

  # --- CURRENT USER PRIVATE RESOURCES (/me/) ---
  /api/v1/me/drafts:
    get:
      operationId: getMyDrafts
      summary: Get current user's draft posts
      tags: [Me]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 0 }
        - name: size
          in: query
          schema: { type: integer, default: 10 }
      responses:
        '200':
          description: Drafts retrieved successfully
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PostPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # --- USER PUBLIC RESOURCES (/users/{id}/) ---
  /api/v1/users/{id}/posts:
    get:
      operationId: getUserPosts
      summary: Get published posts authored by a user
      tags: [Users]
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: page
          in: query
          schema: { type: integer, default: 0 }
        - name: size
          in: query
          schema: { type: integer, default: 10 }
      responses:
        '200':
          description: User posts retrieved successfully
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PostPage' }
        '404': { $ref: '#/components/responses/NotFound' }

  /api/v1/users/{id}/bookmarks:
    get:
      operationId: getUserBookmarks
      summary: Get a user's bookmarked posts
      description: |
        Returns bookmarked posts for the given user. Visibility is controlled by the user's
        `showBookmarks` preference (from UserPreferences). The current user can always access
        their own bookmarks; other users' bookmarks are only visible when `showBookmarks` is true.
      tags: [Users]
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: page
          in: query
          schema: { type: integer, default: 0 }
        - name: size
          in: query
          schema: { type: integer, default: 10 }
      responses:
        '200':
          description: Bookmarked posts retrieved successfully
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PostPage' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  /api/v1/users/{id}/likes:
    get:
      operationId: getUserLikes
      summary: Get posts liked by a user
      description: |
        Returns liked posts for the given user. Visibility is controlled by the user's
        `showLikes` preference (from UserPreferences). The current user can always access
        their own likes; other users' likes are only visible when `showLikes` is true.
      tags: [Users]
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: page
          in: query
          schema: { type: integer, default: 0 }
        - name: size
          in: query
          schema: { type: integer, default: 10 }
      responses:
        '200':
          description: Liked posts retrieved successfully
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PostPage' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # --- TOPICS ---
  /api/v1/topics:
    get:
      operationId: getTopics
      summary: Get all topics grouped by category
      description: Returns all topics with full stats, grouped by category. Used by the Manage Topics page.
      tags: [Topics]
      security: []
      responses:
        '200':
          description: Topics grouped by category
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/TopicCategoryGroup' } }

  /api/v1/topics/by-ids:
    get:
      operationId: getTopicsByIds
      summary: Batch-resolve topic names by ID
      description: |
        Returns minimal `{ id, name }` objects for the given topic IDs — for callers that only
        need to resolve names (e.g. recommendation-service populating subscribed-topic names).
        Unknown IDs are silently omitted; the caller tolerates gaps.

        This is a dedicated sibling of `GET /api/v1/topics` (which returns the full
        category-grouped catalog with stats), mirroring the `GET /api/v1/posts/cards?ids=`
        pattern: one operation, one response shape.
      tags: [Topics]
      security: []
      parameters:
        - name: ids
          in: query
          required: true
          description: Comma-separated list of topic UUIDs, e.g. `ids=id1,id2,id3`. Max 100.
          schema:
            type: array
            items: { type: string, format: uuid }
            minItems: 1
            maxItems: 100
          style: form
          explode: false
      responses:
        '200':
          description: Matching topics retrieved. Unknown IDs are silently omitted.
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Topic' } }
        '400': { $ref: '#/components/responses/BadRequest' }

  /api/v1/topics/search:
    get:
      operationId: searchTopics
      summary: Autocomplete topic search
      tags: [Topics]
      security: []
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string, minLength: 1 }
      responses:
        '200':
          description: Matching topics (max 20)
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/TopicResponse' } }
        '400': { $ref: '#/components/responses/BadRequest' }

  /api/v1/topics/follower-counts:
    post:
      operationId: updateTopicFollowerCounts
      summary: Apply batch follower-count deltas
      description: Increments or decrements follower_count for each named topic. +1 = follow, -1 = unfollow.
      tags: [Topics]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FollowerCountDeltaRequest' }
      responses:
        '204':
          description: Deltas applied
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  /api/v1/topics/trending:
    get:
      operationId: getTrendingTopics
      summary: Get popular topics
      tags: [Topics]
      security: []
      responses:
        '200':
          description: List of trending topics retrieved
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/TopicResponse' } }
        '400': { $ref: '#/components/responses/BadRequest' }

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  responses:
    BadRequest:
      description: Invalid request parameters or body
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: The user is not allowed to perform this action
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: The requested resource was not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }

  schemas:
    # --- SHARED ---
    AuthorSummary:
      type: object
      description: Lightweight author info embedded in posts and comments to avoid extra lookups
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        displayName: { type: string }
        avatarUrl: { type: string, format: uri, nullable: true }
        role: { type: string, enum: [USER, VERIFIED, ADMIN] }
        organisation: { type: string, nullable: true }

    Topic:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }

    # --- POSTS ---
    PostCard:
      type: object
      description: Lightweight post shape for feed cards. Omits full body content.
      properties:
        id: { type: string, format: uuid }
        author: { $ref: '#/components/schemas/AuthorSummary' }
        title: { type: string }
        excerpt: { type: string, nullable: true }
        coverImageUrl: { type: string, format: uri, nullable: true }
        topics: { type: array, items: { $ref: '#/components/schemas/Topic' } }
        readTimeMinutes: { type: integer, nullable: true }
        likeCount: { type: integer }
        commentCount: { type: integer }
        viewCount: { type: integer }
        isLikedByMe: { type: boolean, description: "False when request is unauthenticated" }
        createdAt: { type: string, format: date-time }

    PostRequest:
      type: object
      required: [title, content]
      properties:
        title: { type: string, minLength: 5, maxLength: 100 }
        content: { type: string, minLength: 1, maxLength: 50000 }
        excerpt: { type: string, maxLength: 500, description: "Short preview shown in feed cards. Auto-generated from content if omitted." }
        coverImageUrl: { type: string, format: uri, nullable: true }
        sourceUrl: { type: array, items: { type: string, format: uri } }
        topics: { type: array, items: { type: string }, description: "Topic names. New topics are created automatically." }
        status: { type: string, enum: [DRAFT, PUBLISHED], default: PUBLISHED }

    PostPatchRequest:
      type: object
      description: |
        Partial update payload. All fields are optional — only provided fields are applied.
        Nullable fields (`excerpt`, `coverImageUrl`) can be cleared by sending an explicit null.
      properties:
        title: { type: string, minLength: 5, maxLength: 100 }
        content: { type: string, minLength: 1, maxLength: 50000 }
        excerpt: { type: string, maxLength: 500, nullable: true }
        coverImageUrl: { type: string, format: uri, nullable: true }
        sourceUrl: { type: array, items: { type: string, format: uri } }
        topics: { type: array, items: { type: string }, description: "Topic names. Replaces the post's topics. New topics are created automatically." }
        status: { type: string, enum: [DRAFT, PUBLISHED] }

    PostResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        author: { $ref: '#/components/schemas/AuthorSummary' }
        status: { type: string, enum: [DRAFT, PUBLISHED] }
        title: { type: string }
        excerpt: { type: string, description: "Short preview text for feed cards" }
        content: { type: string, description: "Full post body (Markdown)" }
        coverImageUrl: { type: string, format: uri, nullable: true }
        topics: { type: array, items: { $ref: '#/components/schemas/Topic' } }
        sourceUrl: { type: array, items: { type: string, format: uri } }
        readTimeMinutes: { type: integer }
        likeCount: { type: integer }
        dislikeCount: { type: integer }
        commentCount: { type: integer }
        viewCount: { type: integer }
        saveCount: { type: integer, description: "Number of times this post has been bookmarked" }
        isLikedByMe: { type: boolean, description: "True if the authenticated user liked this post" }
        isDislikedByMe: { type: boolean, description: "True if the authenticated user disliked this post" }
        isBookmarkedByMe: { type: boolean, description: "True if the authenticated user bookmarked this post" }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }

    PostPage:
      type: object
      properties:
        content: { type: array, items: { $ref: '#/components/schemas/PostResponse' } }
        page: { type: integer }
        size: { type: integer }
        totalPages: { type: integer }
        totalElements: { type: integer }

    # --- COMMENTS ---
    CommentRequest:
      type: object
      required: [text]
      properties:
        text: { type: string, minLength: 1, maxLength: 2000 }
        parentId: { type: string, format: uuid, nullable: true, description: "ID of parent comment for replies" }

    CommentResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        author: { $ref: '#/components/schemas/AuthorSummary' }
        text: { type: string }
        likeCount: { type: integer }
        isLikedByMe: { type: boolean }
        createdAt: { type: string, format: date-time }
        replies: { type: array, items: { $ref: '#/components/schemas/CommentResponse' } }

    # --- INTERACTIONS ---
    LikeRequest:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [LIKE, DISLIKE, NONE]
          description: "LIKE or DISLIKE to cast/change reaction. NONE to remove any existing reaction."

    PostLikeResponse:
      type: object
      properties:
        likeCount: { type: integer }
        dislikeCount: { type: integer }
        isLikedByMe: { type: boolean }
        isDislikedByMe: { type: boolean }

    CommentLikeResponse:
      type: object
      properties:
        likeCount: { type: integer }
        isLikedByMe: { type: boolean }

    # --- TOPICS ---
    TopicResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        displayName: { type: string, nullable: true }
        categoryId: { type: string, nullable: true }
        sortOrder: { type: integer }
        totalPostCount: { type: integer }
        postsThisWeek: { type: integer }
        postsPrevWeek: { type: integer }
        activityScore: { type: number, description: "Normalised 0.000–1.000" }
        isHot: { type: boolean }
        followerCount: { type: integer }

    TopicCategoryGroup:
      type: object
      description: A category with all its topics, used by the Manage Topics page.
      properties:
        id: { type: string, description: "Category slug, e.g. 'research'" }
        label: { type: string }
        sortOrder: { type: integer }
        topics:
          type: array
          items: { $ref: '#/components/schemas/TopicResponse' }

    FollowerCountDeltaRequest:
      type: object
      required: [deltas]
      description: Batch follower-count adjustment. Keys are topic name slugs, values are +1 (follow) or -1 (unfollow).
      properties:
        deltas:
          type: object
          additionalProperties: { type: integer }

    # --- ERRORS ---
    ErrorResponse:
      type: object
      properties:
        timestamp: { type: string, format: date-time }
        status: { type: integer }
        error: { type: string }
        message: { type: string }